move entryFromChain to x509util
# Conflicts:
#	internal/hammer/loadtest/workers.go

# Conflicts:
#	internal/scti/handlers.go
