+++ internal/x509util/ct.go
@@ -14,6 +14,7 @@
 package x509util
 
 import (
+	"crypto/sha256"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/asn1"
@@ -21,6 +22,9 @@ import (
 	"fmt"
 	"math/big"
 	"time"
+
+	"github.com/transparency-dev/static-ct/internal/types/rfc6962"
+	"github.com/transparency-dev/trillian-tessera/ctonly"
 )
 
 var (
@@ -190,4 +194,77 @@ func BuildPrecertTBS(tbsData []byte, preIssuer *x509.Certificate) ([]byte, error
 // not exactly 1 CT poison extension present.
 func RemoveCTPoison(tbsData []byte) ([]byte, error) {
 	return BuildPrecertTBS(tbsData, nil)
+}
+
+// EntryFromChain generates an Entry from a chain and timestamp.
+// copied from certificate-transparency-go/serialization.go
+// TODO(phboneff): add tests
+func EntryFromChain(chain []*x509.Certificate, isPrecert bool, timestamp uint64) (*ctonly.Entry, error) {
+	leaf := ctonly.Entry{
+		IsPrecert: isPrecert,
+		Timestamp: timestamp,
+	}
+
+	if len(chain) > 1 {
+		issuersChain := make([][32]byte, len(chain)-1)
+		for i, c := range chain[1:] {
+			issuersChain[i] = sha256.Sum256(c.Raw)
+		}
+		leaf.FingerprintsChain = issuersChain
+	}
+
+	if !isPrecert {
+		leaf.Certificate = chain[0].Raw
+		return &leaf, nil
+	}
+
+	// Pre-certs are more complicated. First, parse the leaf pre-cert and its
+	// putative issuer.
+	if len(chain) < 2 {
+		return nil, fmt.Errorf("no issuer cert available for precert leaf building")
+	}
+	issuer := chain[1]
+	cert := chain[0]
+
+	var preIssuer *x509.Certificate
+	if isPreIssuer(issuer) {
+		// Replace the cert's issuance information with details from the pre-issuer.
+		preIssuer = issuer
+
+		// The issuer of the pre-cert is not going to be the issuer of the final
+		// cert.  Change to use the final issuer's key hash.
+		if len(chain) < 3 {
+			return nil, fmt.Errorf("no issuer cert available for pre-issuer")
+		}
+		issuer = chain[2]
+	}
+
+	// Next, post-process the DER-encoded TBSCertificate, to remove the CT poison
+	// extension and possibly update the issuer field.
+	defangedTBS, err := BuildPrecertTBS(cert.RawTBSCertificate, preIssuer)
+	if err != nil {
+		return nil, fmt.Errorf("failed to remove poison extension: %v", err)
+	}
+
+	leaf.Precertificate = cert.Raw
+	// TODO(phboneff): do we need this?
+	// leaf.PrecertSigningCert = issuer.Raw
+	leaf.Certificate = defangedTBS
+
+	issuerKeyHash := sha256.Sum256(issuer.RawSubjectPublicKeyInfo)
+	leaf.IssuerKeyHash = issuerKeyHash[:]
+	return &leaf, nil
+}
+
+// isPreIssuer indicates whether a certificate is a pre-cert issuer with the specific
+// certificate transparency extended key usage.
+func isPreIssuer(cert *x509.Certificate) bool {
+	// Look for the extension in the Extensions field and not ExtKeyUsage
+	// since crypto/x509 does not recognize this extension as an ExtKeyUsage.
+	for _, ext := range cert.Extensions {
+		if rfc6962.OIDExtKeyUsageCertificateTransparency.Equal(ext.Id) {
+			return true
+		}
+	}
+	return false
 }
